libbind: integrate "netresolv" patches from NetBSD.
libbind development was transferred to the NetBSD project at
http://wiki.netbsd.org/individual-software-releases/netresolv/

There isn't an official release yet, but they provide a set of patches
against the latest libbind release.

* Remove all files we don't use
* Merge the changes to the remaining files
* Add some new files we need
* Move getifaddrs implementation to libnetwork (instead of libbnetapi)
so it can be used by netresolv.

Fixes #8293 : netresolv uses getifaddrs to determine if there is a local
IPv6 address. If there is not, it will not return AAAA records.

+++ headers/posix/arpa/nameser.h
@@ -586,10 +586,10 @@ extern "C" {
 #endif
 
 int		ns_msg_getflag(ns_msg, int);
-u_int	ns_get16(const u_char *);
-u_long	ns_get32(const u_char *);
-void	ns_put16(u_int, u_char *);
-void	ns_put32(u_long, u_char *);
+uint16_t ns_get16(const u_char *);
+uint32_t ns_get32(const u_char *);
+void	ns_put16(uint16_t, u_char *);
+void	ns_put32(uint32_t, u_char *);
 int		ns_initparse(const u_char *, int, ns_msg *);
 int		ns_skiprr(const u_char *, const u_char *, ns_sect, int);
 int		ns_parserr(ns_msg *, ns_sect, int, ns_rr *);
